Added reslicing step to fix error when head was tilted
We added a reslicing step prior to the affine transformation by SPM12.

Previous version resulted in incorrect determination of the face when the image was very tilted towards the back (and maybe the front). SPM12 was unable to return a tilted native-to-MNI transformation matrix. As a result, the face was located in a too low position, covering only the bottom part of the cheeks and the chin.

Current version adds a reslicing step after locating the anterior and posterior commissures, so both landmarks lay in a completely horizontal plane. SSPM12 is able to correctly transform the resliced MRI into MNI space, and this transformation is combined with the native-to-AC/PC to generate a correct native-to-MNI transformation.